oldamtgui/requirements.txt
pyaudio
PySide6 
omegaconf
h5py
pytz
torchaudio
matplotlib
scipy
qdarkstyle
coloredlogs
pandas